// types.ts
// types.ts
export type RootStackParamList = {
    Home: undefined;
    Welcome: undefined;
    Landscaping: { courseId: string; price: number};
    FirstAid: { courseId: string; price: number};
    Sewing: { courseId: string; price: number};
    LifeSkills: { courseId: string; price: number};
    ChildMinding: { courseId: string; price: number};
    Cooking: { courseId: string; price: number};
    GardenMaintenance: { courseId: string; price: number};
    Quote: undefined;
    LongCourses: undefined;
    ShortCourses: undefined;
    ContactUs: undefined;
  };
  